# public/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp Business SaaS API",
    "description": "API REST para integración con la plataforma de WhatsApp Business SaaS. Esta API permite gestionar contactos, campañas y mensajes de WhatsApp.\n\n## Autenticación\n\nTodas las rutas protegidas requieren un token Bearer. Obtén tu token usando el endpoint `/api/v1/auth/login`.\n\n## Rate Limiting\n\nLa API tiene límites de uso según tu plan de suscripción:\n- **Free**: 100 requests/minuto\n- **Basic**: 500 requests/minuto\n- **Professional**: 2000 requests/minuto\n- **Enterprise**: Sin límites\n\n## Códigos de Estado\n\n- `200` - Éxito\n- `201` - Recurso creado\n- `400` - Error de validación\n- `401` - No autenticado\n- `403` - No autorizado / Límite alcanzado\n- `404` - Recurso no encontrado\n- `500` - Error del servidor",
    "version": "1.0.0",
    "contact": {
      "name": "Soporte API",
      "email": "api@whatsapp-saas.com"
    }
  },
  "servers": [
    {
      "url": "https://forestgreen-hamster-482261.hostingersite.com/api/v1",
      "description": "Producción"
    },
    {
      "url": "/api/v1",
      "description": "API v1 (Relativo)"
    },
    {
      "url": "http://127.0.0.1:8000/api/v1",
      "description": "Desarrollo Local (Laravel Serve)"
    },
    {
      "url": "http://whatsapp.local/api/v1",
      "description": "Desarrollo Local (Apache VirtualHost)"
    },
    {
      "url": "http://localhost/mexico/api/v1",
...
    }
  ],
...
  "paths": {
...
    "/messages/send-text": {
      "post": {
        "tags": ["Messages"],
        "summary": "Enviar mensaje de texto",
        "description": "Envía un mensaje de texto a un contacto",
        "operationId": "sendTextMessage",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["contact_id", "message"],
                "properties": {
                  "contact_id": {
                    "type": "integer",
                    "description": "ID del contacto"
                  },
                  "message": {
                    "type": "string",
                    "maxLength": 4096,
                    "description": "Contenido del mensaje"
                  },
                  "waba_account_id": {
                    "type": "integer",
                    "description": "ID de la cuenta WABA (opcional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mensaje enviado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMessageResponse"
                }
              }
            }
          },
          "403": {
            "description": "Límite de mensajes alcanzado"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token"
      }
    },
    "schemas": {
...
      "SendMessageResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "message": {"type": "string"},
          "data": {
            "type": "object",
            "properties": {
              "message_id": {"type": "integer"},
              "whatsapp_message_id": {"type": "string"},
              "status": {"type": "string"}
            }
          }
        }
      },
...
    }
  }
}